+++ spec/kettle/dev/release_cli_spec.rb
@@ -2,8 +2,6 @@
 
 # rubocop:disable RSpec/MultipleExpectations, RSpec/MessageSpies, RSpec/StubbedMock, RSpec/ReceiveMessages
 
-require "kettle/dev/ci_monitor"
-
 RSpec.describe Kettle::Dev::ReleaseCLI do
   let(:ci_helpers) { Kettle::Dev::CIHelpers }
   let(:cli) { described_class.new }